requirements.txt
fastapi
pycryptodome
gmssl